🚑️ Fix project build 3

+++ filekit-core/src/appleMain/kotlin/io/github/vinceglb/filekit/utils/NSDataExt.kt
@@ -21,6 +21,6 @@ public fun NSData.toByteArray(): ByteArray = let { nsData ->
 public fun ByteArray.toNSData(): NSData = usePinned {
     NSData.create(
         bytes = it.addressOf(0),
-        length = this.size.toUInt(),
+        length = this.size.toULong(),
     )
 }
